add visualization for exercise module
