+++ src/main/frontend/view/exercise/visualization/ExerciseVisualBuilder.java
@@ -1,0 +1,4 @@
+package main.frontend.view.exercise.visualization;
+
+public class ExerciseVisualBuilder {
+}
