+++ src/main/frontend/view/exercise/visualization/ExerciseVisualDirector.java
@@ -1,0 +1,4 @@
+package main.frontend.view.exercise.visualization;
+
+public class ExerciseVisualDirector {
+}
