+++ src/main/frontend/view/exercise/visualization/ExerciseVisualPage.java
@@ -1,0 +1,20 @@
+package main.frontend.view.exercise.visualization;
+
+import main.backend.common.Result;
+import main.backend.exercise.IExerciseController;
+import main.backend.exercise.entity.Exercise;
+import main.backend.exercise.impl.ExerciseController;
+import main.backend.user.entity.User;
+import main.frontend.custom.dropdown.AutoComboBox;
+import main.frontend.custom.entry.NfEntry;
+import main.frontend.custom.table.PaginationTable;
+import main.frontend.common.Content;
+
+import javax.swing.*;
+
+public class ExerciseVisualPage extends Content {
+    @Override
+    public String showContent(JPanel content) {
+        return null;
+    }
+}
